requirements.txt
pytest==6.2.3
python-dotenv==1.0.0
selenium==4.1.0
urllib3==1.26.4